+++ client/private/js/src/main.js
@@ -34,6 +34,7 @@ whenDomReady( () => {
   // GlobalEvents bindings:
   globalEvents.bindEvent( 'click', ( e ) => { contextMenu.outsideClickGlobalEvent( e ); } );
   globalEvents.bindEvent( 'click', ( e ) => { startMenuManager.outsideClickGlobalEvent( e ); } );
+  globalEvents.bindEvent( 'click', ( e ) => { document.getElementById( 'click-audio' ).play(); } );
   globalEvents.init();
 
   //console.debug( 'FS V2:', fileSystem.____fsv2 );