+++ client/private/js/src/systemApps/profiles/exploreProfiles/exploreProfiles.model.js
@@ -9,6 +9,16 @@
 
 class ExploreProfilesModel {
   constructor() {
+    this.targetWindow;
+  }
 
+  async getUsersPage( lastId = 0 ) {
+    const res = await HttpClient.get( `${API_ROOT_PATH}users/last-logged-in?lastId=${lastId}&limit=10` );
+    if ( !res.ok() ) {
+      Notifications.errorToast( 'There has been an error while getting the users page. Please try again later.' );
+      return false;
+    }
+
+    return await res.json();
   }
 }