+++ client/private/js/src/systemApps/profiles/exploreProfiles/exploreProfiles.view.js
@@ -9,6 +9,20 @@
 
 class ExploreProfilesView {
   constructor() {
+  }
+
+  get allUserCards() { return DomUtils.get( '#' + this.model.targetWindow + '.user-card' ); }
+
+  injectCards( target, users ) {
+    const container = document.getElementById( 'cntnt_' + target );
+    container.innerHTML = '';
+
+    for ( let i = 0; i < userCards.length; ++i ) {
+      container.innerHTML += ExploreProfilesTemplates.userCard();
+    }
+  }
+
+  injectProfile( target, content ) {
 
   }
 }