+++ client/wwwroot/js/src_portfolio-os/portfolio-os_auth.header.js
@@ -1,2 +1,3 @@
 ﻿// %import<<GH 'js-cookie/js-cookie/master/src/js.cookie.js'
+// %import<<GH 'ckknight/random-js/lib/random.min.js'
 // @import 'portfolio-os_auth'